moves admin pages to admin namspace
