+++ src/Cogitatio/Pages/Admin/AdminAddPost.razor
@@ -1,4 +1,4 @@
-﻿@page "/AdminAddPost"
+﻿@page "/a/AddPost"
 
 <h3>Create a new post</h3>
 @if (AdminUserState.IsAdmin)
