+++ src/Cogitatio/Pages/Admin/AdminContact.razor
@@ -1,4 +1,4 @@
-﻿@page "/AdminContact"
+﻿@page "/a/Contact"
 <h3>Contact Admin</h3>
 
 @if (AdminUserState.IsAdmin == true)
