+++ src/Cogitatio/Pages/Admin/AdminDiag.razor
@@ -1,4 +1,4 @@
-﻿@page "/AdminDiag"
+﻿@page "/a/Diag"
 
 <h3>Cogitatio Configuration</h3>
