+++ src/Cogitatio/Pages/Admin/AdminDiag.razor.css
