+++ src/Cogitatio/Pages/Admin/AdminEditPost.razor
@@ -1,5 +1,5 @@
-﻿@page "/AdminEditPost"
-@page "/AdminEditPost/{Slug}"
+﻿@page "/a/EditPost"
+@page "/a/EditPost/{Slug}"
 <h3>Editing Post</h3>
 
 @if (AdminUserState.IsAdmin)
