+++ src/Cogitatio/Pages/Admin/AdminEditPost.razor.css
