+++ src/Cogitatio/Pages/Admin/AdminHome.razor
@@ -44,27 +44,27 @@
             <h5 class="card-title text-center">Admin Actions</h5>
             <div class="d-flex flex-wrap justify-content-center">
                 <div class="action-item">
-                    <NavLink href="AdminDiag" class="btn btn-outline-primary">Diag</NavLink>
+                    <NavLink href="/a/Diag" class="btn btn-outline-primary">Diag</NavLink>
                     <span class="action-description">View diagnostic information</span>
                 </div>
                 <div class="action-item">
-                    <NavLink href="AdminSettings" class="btn btn-outline-primary">Settings</NavLink>
+                    <NavLink href="/a/Settings" class="btn btn-outline-primary">Settings</NavLink>
                     <span class="action-description">Site Configuration</span>
                 </div>
                 <div class="action-item">
-                    <NavLink href="AdminContact" class="btn btn-outline-primary">Contacts</NavLink>
+                    <NavLink href="/a/Contact" class="btn btn-outline-primary">Contacts</NavLink>
                     <span class="action-description">View contacts</span>
                 </div>
                 <div class="action-item">
-                    <NavLink href="AdminAddPost" class="btn btn-outline-primary">Add New Post</NavLink>
+                    <NavLink href="/a/AddPost" class="btn btn-outline-primary">Add New Post</NavLink>
                     <span class="action-description">Create a new post</span>
                 </div>
                 <div class="action-item">
-                    <NavLink href="AdminEditPost" class="btn btn-outline-primary">Edit Post</NavLink>
+                    <NavLink href="/a/EditPost" class="btn btn-outline-primary">Edit Post</NavLink>
                     <span class="action-description">Edit a post</span>
                 </div>
                 <div class="action-item">
-                    <NavLink href="TagEditor" class="btn btn-outline-primary">Edit Tags</NavLink>
+                    <NavLink href="/a/TagEditor" class="btn btn-outline-primary">Edit Tags</NavLink>
                     <span class="action-description">Quickly edit tags for multiple posts</span>
                 </div>
                 <div class="action-item">
