+++ src/Cogitatio/Pages/Admin/AdminSettings.razor
@@ -1,4 +1,4 @@
-﻿@page "/AdminSettings"
+﻿@page "/a/Settings"
 @using TinyMCE.Blazor
 
 <h3>Admin Settings</h3>
