+++ src/Cogitatio/Pages/Admin/AdminSettings.razor.css
