+++ src/Cogitatio/Pages/Admin/TagEditor.razor
@@ -1,4 +1,4 @@
-﻿@page "/TagEditor"
+﻿@page "/a/TagEditor"
 <h3>Tag Editor</h3>
 
 @* Tag summary cards: responsive, multiple per row *@
